Resvidex pipeline

Wrapper function that performs all the recommended step. This includes

kcounter

# Count Kmers and normalize data

prediction_caller

Run prediction pipeline

quality_control

Check Quality of the input data and classification results

quality_filter

For sequences with any FLAG present set result Clade to LowQuality

Usage

classify(inputFile, model, QC_value = 0.4, Length_value = 0.5, N_value = 2)

Arguments

inputFile

string path relative to the working directory of the input file. Must be in fasta format.

model

A random forest classification model. Must be FULL_GENOME. See FULL_GENOME.

QC_value

numeric value from 0.3 to 1. (default = 0.4)

Length_value

numeric value from 0 to 1. Proportion of difference to the expected sequence length. (default = 0.5)

N_value

numeric value from 0 to 100. Percentage of acceptable ambiguous bases. (default = 2)

Value

a dataframe

Examples

file_path<-system.file("extdata","test_dataset.fasta",package="resvidex")

Classification<-classify(inputFile=file_path,model=FULL_GENOME)

#the variable Classification has now all the sequences classified
Classification
#>                                                                            Label
#> 1 KM360090.1 Respiratory syncytial virus isolate RSV Memphis-37, complete genome
#>   Clade Probability Length Length_QC N N_QC Probability_QC
#> 1 A.2.1       0.785  15147      TRUE 0 TRUE           TRUE

if (FALSE) {
#if you want to export the prediction
utils::write.csv2(Classification,"Classification_file.csv")
}