prediction_caller

Performs the prediction and computes probability values.

Usage

prediction_caller(NormalizedData, model, QC_unknown = 0.3)

Arguments

NormalizedData

A list of 3 elements: normalized k-mer counts, genome length and contents of undefined bases.Produced by the kcounter() function

model

A random forest classification model. Must be FULL_GENOME. See FULL_GENOME.

QC_unknown

numeric value from 0 to 1. Stringent filter, do not classify below this probability score (default = 0.2)

Value

Data.frame with the classification results and quality checks. The output has the following properties:

Each row corresponds to one sample sequence.

Label is the name of the sequence.

Clade is the corresponding prediction.

Probability is the proportions of trees that agreed with the Clade result. Values between 0 to 1.

Probability_QC a logical value. If TRUE the sequence passed the quality filter for probability. Run quality_control() to fill this column.

Length Sequence length.

Length_QC a logical value. If TRUE the sequence passed the quality filter for length.Run quality_control() to fill this column

N proportions of undefined bases in the sequence. The lower the better.

N_QC a logical value. If TRUE the sequence passed the quality filter for undefined bases. Run quality_control() to fill this column

Examples

file_path<-system.file("extdata","test_dataset.fasta",package="resvidex")

sequence<-ape::read.FASTA(file_path,type = "DNA")

NormalizedData <- kcounter(SequenceData=sequence,model=FULL_GENOME)
#> Error in kcounter(SequenceData = sequence, model = FULL_GENOME): could not find function "kcounter"

PredictedData <- prediction_caller(NormalizedData=NormalizedData,model=FULL_GENOME)
#> Error in prediction_caller(NormalizedData = NormalizedData, model = FULL_GENOME): could not find function "prediction_caller"